add full comments functional

+++ app/Http/Controllers/Post/Comment/DestroyController.php
@@ -5,14 +5,38 @@ namespace App\Http\Controllers\Post\Comment;
 use App\Http\Controllers\Controller;
 use App\Models\PostComment;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\DB;
 
 class DestroyController extends Controller
 {
     public function __invoke(PostComment $comment)
     {
 
-        $comment->delete();
-        return response()->json(['message' => 'deleted successful']);
+        $firstCondition = $comment->user->id === auth()->user()->id;
+        $secondCondition = $comment->post->user->id === auth()->user()->id;
+
+        if($firstCondition || $secondCondition)
+        {
+            try {
+                DB::beginTransaction();
+
+                $comments = PostComment::where('comment_id', $comment->id)->get();
+
+                foreach ($comments as $comm)
+                {
+                    $comm->delete();
+                }
+
+                $comment->delete();
+                DB::commit();
+                return response()->json(['message' => 'deleted successful']);
+            } catch (\Exception $e) {
+                DB::rollBack();
+            }
+
+        }
+
+
 
     }
 }
